feat: Add BDD Tests for Ducks
Use the Ginkgo framework to add behavior-driven automated tests. This
approach focuses on testing behaviors that deliver value, as opposed
to standard unit tests which are not opinionated. In my experience,
developers tend to focus on implementation details with standard unit
tests, making the overall test suite challenging to update when new
features are introduced.