feat: Initial Buggy Ducks Implementation
This is an intentionally buggy initial implementation of the Duck
objects and behaviors. To facilitate testing, the Duck structs each
have a `Writer` field of type `io.StringWriter`. The "zero value"
should default to writing to `os.Stdout` (go proverb: "Make the zero
value useful"). A new Ginkgo test suite verifies the behavior of each
Duck type.

There is a lot here that is wrong - an exercise for the "reader" is to
correct these errors.

Signed-off-by: Adam B Kaplan <adam@adambkaplan.com>