feat: Allow Behaviors to Change for All Ducks
Per the spec in "Head First Design Patterns," all ducks can change
their flying and quacking behaviors. With the interface update, it made
sense to introduce a "base" GenericDuck struct that the other Duck
structs embed. This makes it easier to add other duck types that change
their default behaviors and/or displays.

Signed-off-by: Adam B Kaplan <adam@adambkaplan.com>