refactor: Concurrent Observer
Refactor the initial Observer implementation so that the Subject sends
concurrent notifications. Unlike Java, Go has built in primitives
designed to natively manage parallel execution of code. In this
implementation, the "Subject" is responsible for managing the
goroutines, channels, and other aspects of running concurrent code.
The Observers do not need to know that their code is being run in
separate goroutines.

Adding concurrency required updates to the tests, which assumed serial
execution. Items that were procedurally true now must be _eventually_
true, and black box "harness" code needed to add delays to give the
displays time to write their content.

Signed-off-by: Adam B Kaplan <adam@adambkaplan.com>